// typings.d.ts
export interface Zakuski{
    _id: string;
    title:string;
    description:string;
    mainImage:{
        asset:{
            url:string
        }
    }
    slug:{
        current:string;
    }
    body:[object]
}

export interface Hlqb{
    _id: string;
    title:string;
    description:string;
    mainImage:{
        asset:{
            url:string
        }
    }
    slug:{
        current:string;
    }
    body:[object]
}

export interface Pogachi{
    _id: string;
    title:string;
    description:string;
    mainImage:{
        asset:{
            url:string
        }
    }
    slug:{
        current:string;
    }
    body:[object]
}

export interface Kozunaci{
    _id: string;
    title:string;
    description:string;
    mainImage:{
        asset:{
            url:string
        }
    }
    slug:{
        current:string;
    }
    body:[object]
}

export interface Sladki{
    _id: string;
    title:string;
    description:string;
    mainImage:{
        asset:{
            url:string
        }
    }
    slug:{
        current:string;
    }
    body:[object]
}

export interface Napitki{
    _id: string;
    title:string;
    description:string;
    mainImage:{
        asset:{
            url:string
        }
    }
    slug:{
        current:string;
    }
    body:[object]
}